Move some functions from util to hw
