+++ pyanaconda/core/hw.py
@@ -17,6 +17,7 @@
 # along with this program.  If not, see <http://www.gnu.org/licenses/>.
 #
 import blivet.arch
+from pyanaconda.core.util import execWithCapture
 
 from pyanaconda.anaconda_loggers import get_module_logger
 log = get_module_logger(__name__)
@@ -68,3 +69,61 @@ def total_memory():
 
         log.error("MemTotal: line not found in /proc/meminfo")
         raise RuntimeError("MemTotal: line not found in /proc/meminfo")
+
+
+def detect_virtualized_platform():
+    """Detect execution in a virtualized environment.
+
+    This runs systemd-detect-virt and, if the result is not 'none',
+    it returns an id of the detected virtualization technology.
+
+    Otherwise, it returns None.
+
+    :return: a virtualization technology identifier or None
+    """
+    try:
+        platform = execWithCapture("systemd-detect-virt", []).strip()
+    except (OSError, AttributeError):
+        return None
+
+    if platform == "none":
+        return None
+
+    return platform
+
+
+def is_smt_enabled():
+    """Is Simultaneous Multithreading (SMT) enabled?
+
+    :return: True or False
+    """
+    from pyanaconda.flags import flags
+    from pyanaconda.core.configuration.anaconda import conf
+
+    if flags.automatedInstall \
+            or not conf.target.is_hardware \
+            or not conf.system.can_detect_enabled_smt:
+        log.info("Skipping detection of SMT.")
+        return False
+
+    try:
+        return int(open("/sys/devices/system/cpu/smt/active").read()) == 1
+    except (OSError, ValueError):
+        log.warning("Failed to detect SMT.")
+        return False
+
+
+def is_lpae_available():
+    """Is LPAE available?
+
+    :return: True of False
+    """
+    if not blivet.arch.is_arm():
+        return False
+
+    with open("/proc/cpuinfo", "r") as f:
+        for line in f:
+            if line.startswith("Features") and "lpae" in line.split():
+                return True
+
+    return False
